# docs/api/openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    Page_UserOut_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/UserOut'
          title: Items
          type: array
        page:
          title: Page
          type: integer
        page_size:
          title: Page Size
          type: integer
        total:
          title: Total
          type: integer
      required:
      - items
      - page
      - page_size
      - total
      title: Page[UserOut]
      type: object
...
    UserOut:
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        email:
          format: email
          title: Email
          type: string
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
        id:
          format: uuid
          title: Id
          type: string
        last_login_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Last Login At
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        roles:
          items:
            type: string
          title: Roles
          type: array
        status:
          title: Status
          type: string
        two_factor_enabled:
          title: Two Factor Enabled
          type: boolean
      required:
      - id
      - email
      - status
      - two_factor_enabled
      - created_at
      title: UserOut
      type: object
    UserRoleAssignIn:
      properties:
        role_code:
          title: Role Code
          type: string
      required:
      - role_code
      title: UserRoleAssignIn
      type: object
    UserUpdateIn:
      properties:
        full_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Full Name
        phone:
          anyOf:
          - type: string
          - type: 'null'
          title: Phone
        status:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      title: UserUpdateIn
      type: object
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: Kaya BOS
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /api/v1/users:
    get:
      operationId: list_users_api_v1_users_get
      parameters:
      - in: query
        name: page
        required: false
        schema:
          default: 1
          minimum: 1
          title: Page
          type: integer
      - in: query
        name: page_size
        required: false
        schema:
          default: 25
          maximum: 200
          minimum: 1
          title: Page Size
          type: integer
      - in: query
        name: q
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Q
      - in: query
        name: status
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_UserOut_'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Users
      tags:
      - users
  /api/v1/users/{user_id}:
    get:
      operationId: get_user_api_v1_users__user_id__get
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          format: uuid
          title: User Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get User
      tags:
      - users
    patch:
      operationId: update_user_api_v1_users__user_id__patch
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          format: uuid
          title: User Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Update User
      tags:
      - users
  /api/v1/users/{user_id}/roles:
    post:
      operationId: assign_role_api_v1_users__user_id__roles_post
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          format: uuid
          title: User Id
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRoleAssignIn'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Assign Role
      tags:
      - users
  /api/v1/users/{user_id}/roles/{role_code}:
    delete:
      operationId: revoke_role_api_v1_users__user_id__roles__role_code__delete
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          format: uuid
          title: User Id
          type: string
      - in: path
        name: role_code
        required: true
        schema:
          title: Role Code
          type: string
      - in: header
        name: authorization
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserOut'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Revoke Role
      tags:
      - users
...